Create deploy-docker.yaml
This is the base file that has been tested to work. This will deploy docker and compose and create a user for regular access and add it to the docker group. You will still need to create a password for the user using the root login before you are able to login with that created user.